+++ src/test/scala/org/broadinstitute/dsde/workbench/leonardo/dao/MockJupyterDAO.scala
@@ -4,7 +4,7 @@ import org.broadinstitute.dsde.workbench.model.google.GoogleProject
 
 import scala.concurrent.Future
 
-class MockJupyterDAO extends ToolDAO{
+object MockJupyterDAO extends ToolDAO{
 
   override def isProxyAvailable(googleProject: GoogleProject, clusterName: ClusterName): Future[Boolean] = Future.successful(true)
