+++ src/test/scala/org/broadinstitute/dsde/workbench/leonardo/dao/MockRStudioDAO.scala
@@ -5,7 +5,7 @@ import org.broadinstitute.dsde.workbench.model.google.GoogleProject
 
 import scala.concurrent.Future
 
-class MockRStudioDAO extends ToolDAO{
+object MockRStudioDAO extends ToolDAO{
 
   override def isProxyAvailable(googleProject: GoogleProject, clusterName: ClusterName): Future[Boolean] = Future.successful(true)
